feat: add aria live in dg footer
