+++ packages/pluggableWidgets/datagrid-web/src/components/WidgetFooter.tsx
@@ -1,6 +1,7 @@
 import { If } from "@mendix/widget-plugin-component-kit/If";
 import { observer } from "mobx-react-lite";
 import { ReactElement } from "react";
+import { SelectionAriaLive } from "../features/selection-counter/SelectionAriaLive";
 import { SelectionCounter } from "../features/selection-counter/SelectionCounter";
 import { useSelectionCounterViewModel } from "../features/selection-counter/injection-hooks";
 import { useCustomPagination, usePaginationConfig, usePaginationVM, useTexts } from "../model/hooks/injection-hooks";
@@ -25,6 +26,9 @@ export const WidgetFooter = observer(function WidgetFooter(): ReactElement | nul
 
     return (
         <div className="widget-datagrid-footer table-footer">
+            <If condition={config.selectionEnabled}>
+                <SelectionAriaLive />
+            </If>
             <div className="widget-datagrid-paging-bottom">
                 <div className="widget-datagrid-pb-start">
                     <If condition={selectionCounterVM.isBottomCounterVisible}>
