+++ packages/pluggableWidgets/datagrid-web/src/features/selection-counter/injection-hooks.ts
@@ -1,4 +1,5 @@
 import { createInjectionHooks } from "brandi-react";
-import { DG_TOKENS } from "../../model/tokens";
+import { CORE_TOKENS, DG_TOKENS } from "../../model/tokens";
 
 export const [useSelectionCounterViewModel] = createInjectionHooks(DG_TOKENS.selectionCounterVM);
+export const [useSelectionCounterTexts] = createInjectionHooks(CORE_TOKENS.selection.selectedCounterTextsStore);
