+++ packages/pluggableWidgets/datagrid-web/src/model/tokens.ts
@@ -5,6 +5,7 @@ import { ClickActionHelper } from "@mendix/widget-plugin-grid/helpers/ClickActio
 import { FocusTargetController } from "@mendix/widget-plugin-grid/keyboard-navigation/FocusTargetController";
 import { VirtualGridLayout } from "@mendix/widget-plugin-grid/keyboard-navigation/VirtualGridLayout";
 import {
+    ObservableSelectAllTexts,
     QueryService,
     SelectActionsService,
     SelectAllService,
@@ -20,11 +21,7 @@ import {
     SetPageAction
 } from "@mendix/widget-plugin-grid/pagination/main";
 import { SelectAllFeature } from "@mendix/widget-plugin-grid/select-all/select-all.feature";
-import {
-    BarStore,
-    ObservableSelectAllTexts,
-    SelectAllEvents
-} from "@mendix/widget-plugin-grid/select-all/select-all.model";
+import { BarStore, SelectAllEvents } from "@mendix/widget-plugin-grid/select-all/select-all.model";
 import { SelectionCounterViewModel } from "@mendix/widget-plugin-grid/selection-counter/SelectionCounter.viewModel-atoms";
 import { ComputedAtom, DerivedPropsGate, Emitter } from "@mendix/widget-plugin-mobx-kit/main";
 import { token } from "brandi";
@@ -81,7 +78,8 @@ export const CORE_TOKENS = {
         selectedCounterTextsStore: token<{
             clearSelectionButtonLabel: string;
             selectedCountText: string;
-        }>("@store:selectedCounterTextsStore")
+        }>("@store:selectedCounterTextsStore"),
+        selectAllTexts: token<ObservableSelectAllTexts>("@store:SelectAllTexts")
     },
 
     setupService: token<DatagridSetupService>("DatagridSetupService"),
@@ -157,7 +155,6 @@ export const SA_TOKENS = {
     emitter: token<Emitter<SelectAllEvents>>("SelectAllEmitter"),
     gate: token<DerivedPropsGate<MainGateProps>>("MainGateForSelectAllContainer"),
     progressService: token<TaskProgressService>("SelectAllProgressService"),
-    selectAllTextsStore: token<ObservableSelectAllTexts>("SelectAllTextsStore"),
     selectAllBarVM: token<SelectAllBarViewModel>("SelectAllBarViewModel"),
     selectAllService: token<SelectAllService>("SelectAllService"),
     selectionDialogVM: token<SelectionProgressDialogViewModel>("SelectionProgressDialogViewModel"),
